requirements.txt
numpy>=1.21.0
scikit-learn>=1.0.0
pandas>=1.5.0
matplotlib>=3.5.0
seaborn>=0.11.0
tqdm>=4.65.0
faiss-cpu>=1.7.0
sentence-transformers>=2.2.0
torch>=1.9.0
transformers>=4.20.0
sqlalchemy>=1.4.0